comment建表语句添加type字段

+++ reshare-model/src/main/java/com/xktime/model/article/pos/CrawlerArticle.java
@@ -20,6 +20,6 @@ public class CrawlerArticle extends BaseVerifyArticle {
     private Date publishTime;
     private String url;
     private String origin;
-    private int status;
+    private byte status;
 
 }
